Home

This document provides the documentation for the package StochTools.jl. StochTools.jl provides a collection of structures and functions for use in simulating the Gillespie SSA and related hybrid methods.

This package was written by Joshua C. Kynaston of the University of Bath.

Structures

Stochasticity.GillespieParams — Type

GillespieParams(tf::Float64, Δt::Float64, IC::Tuple{Vararg{Int64}}, λ::Tuple, F::Function, repn::Int64)

Holds the parameters for a run of the Gillespie Direct Algorithm.

Model Parameters

tf :: Float64: Final runtime

Δt :: Float64: Recording time step

K :: Int64: Number of chemical species

IC :: Tuple{Vararg{Int64}}

λ :: Tuple{Vararg{Float64}}: Vector of rate parameters

F :: Function: Function to calculate propensities at runtime

R :: Matrix{Int64}: Stoichiometry matrix

repn :: Int64: Number of repeats

Methods

Stochasticity.record_state! — Function

record_state!()

Stochasticity.sample_dist — Method

sample_dist(a::Union{AbstractArray{T}, Tuple{Vararg{T}}}) where T <: Real

Samples from a discrete probability distribution.

Arguments

a: Array of probabilities