Home

This document provides the documentation for the package StochTools.jl

This package was written by Joshua C. Kynaston of the University of Bath.

Structures

StochTools.Gillespie_Model — Type

Gillespie_Model(...)

A mutable struct for storing the parameters and results of a run of the direct Gillespie SSA.

Instantiating this struct requires input of the following mandatory model parameters that define the basic model behaviour. Specifying this set of parameters is sufficient to run all subsequent methods.

Model Parameters

tf::Float64: Final runtime

Δt::Float64: Recording time step

K::Int64: Number of chemical species

λ::Vector{Union{Int64,Float64}}: Vector of rate parameters

Initialising

This struct is defined via the Parameters.jl package with the @with_kw flag. As such, initialisation requires that all mandatory parameters be passed as name-value pairs. For example,

    model = Gllespie_Model(tf=10, Δt=1e-3, K=5,...)

The remaining fields that are initialised in this struct are either derived parameters/data structures, listed below as derived fields, or pre-allocated fields containing matrices for storing additional model data, listed as auxilliary fields.

Derived Fields

t::Float64 = 0.0: Current model time.

Nt::Float64 = floor(tf/Δt) + 1: Number of time steps.

tp::Int64 = 0: Index of previous time step.

Structural Fields

state::Matrix{Int64} = Matrix{Int64}(undef, K, 1): Stores the current state. Initialised with dimensions (K,1)

stoich::Union{Matrix{Int64},Vector{Int64}} = Matrix{Int64}(undef, K, length(λ)): Stoichiometric matrix.

rec::Matrix{Int64} = Matrix{Int64}(undef, K, Nt): Stores trajectories.

α:: Vector{Float64} = Vector{Float64}(undef, length(λ)): Vector of calculated propensities.

Methods

StochTools.sample_reaction — Method

sample_reaction(::Vector{Float64})

Samples from a discrete probability distribution.

Arguments

a::Vector{Any}: Vector of probabilities

StochTools.sample_reaction — Method

sample_reaction(::Vector{Float64},::Float64)

Samples from an unnormalised discrete distribution.

Arguments

a::Vector{Float64}: Vector of weights

a₀::Float64: Sum of weights